requirements.txt
numpy
matplotlib
pygltflib
tqdm